backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
reportlab==4.2.5
transformers==4.44.2
torch==2.7.1
